requirements.txt
numpy==1.24.1
tqdm==4.65.0
omegaconf==2.3.0
albumentationsx==2.0.10
scipy==1.15.3
